% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensuSV.r
\name{format_delly}
\alias{format_delly}
\title{Title: Format Delly}
\usage{
format_delly(delly)
}
\arguments{
\item{delly}{Data frame with delly output}
}
\description{
This function extract required fields from output vcf
}
