% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensuSV.r
\name{format_gridss}
\alias{format_gridss}
\title{Title: Format GRIDSS2}
\usage{
format_gridss(gridss)
}
\arguments{
\item{gridss}{Data frame with GRIDSS2 output}
}
\description{
This function extract required fields from output vcf
}
