% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensuSV.r
\name{format_manta}
\alias{format_manta}
\title{Title: Format Manta}
\usage{
format_manta(manta)
}
\arguments{
\item{manta}{Data frame with manta output}
}
\description{
This function extract required fields from output vcf
}
