% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensuSV.r
\name{format_svaba}
\alias{format_svaba}
\title{Title: Format svaba}
\usage{
format_svaba(svaba)
}
\arguments{
\item{svaba}{Data frame with svaba output}
}
\description{
This function extract required fields from output vcf
}
