% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensuSV.r
\name{make_sv_consensus}
\alias{make_sv_consensus}
\title{Title: Make consensus from different callers}
\usage{
make_sv_consensus(df, max_dist = 200)
}
\arguments{
\item{df}{concatenate bedpe format SV call results}

\item{max_dist}{amx distance to consider to events (default 200bp)}
}
\description{
This function to create consensus from different callers
}
