% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensuSV.r
\name{parse_strands_from_alt}
\alias{parse_strands_from_alt}
\title{Title: Parse strand from ALT/SVTYPE field}
\usage{
parse_strands_from_alt(alt, ref)
}
\arguments{
\item{alt}{is the ALT field which can have shorthand notation (e.g. \if{html}{\out{<DEL>}}), square brackets or genomic sequence}

\item{ref}{REF field which is the reference}
}
\description{
This function infer strand from ALT field
}
