% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensuSV.r
\name{rename_SVclass}
\alias{rename_SVclass}
\title{Title: Format and harmonize SVtype field}
\usage{
rename_SVclass(SVTYPE, chr1, chr2, strand1 = NA, strand2 = NA)
}
\arguments{
\item{SVTYPE}{from formatted function}

\item{chr1}{first chromosome coordinate}

\item{chr2}{second chromosome coordinate}

\item{strand1}{of first BND}

\item{strand2}{of second BND}
}
\description{
This function homogenize SVTYPE field
}
